fix(query): extract QueryEntityGraphFetchingOptimizer in separate class
